refactor: config moved
