+++ inspector-runtime/src/main/java/com/uhufor/inspector/config/ConfigProvider.kt
@@ -1,4 +1,4 @@
-package com.uhufor.inspector
+package com.uhufor.inspector.config
 
 interface ConfigProvider {
     fun getConfig(): Config
